"Added room quantity, host ID, and image upload functionality to room controller and service"

+++ src/controllers/room-controller.js
@@ -1,30 +1,50 @@
 const createError = require("../utils/createError");
 const roomService = require("../service/roomService");
 const { v4: uuidv4 } = require("uuid");
+const cloudUpload = require("../utils/cloudUpload");
 
 exports.createRoom = async (req, res, next) => {
     try {
         const {
             name,
+            quantity,
             type,
             price,
-            hostId
         } = req.body;
 
-        if (!name || !type || !price || !hostId) {
-            return next(createError(400, 'Name, type, price, and host ID are required.'));
+        if (!name || !quantity || !type || !price) {
+            return next(createError(400, 'Name, quantity, type, price, and host ID are required.'));
         }
 
+        const imagexPromiseArray = req.files.map((file) => {
+            return cloudUpload(file.path)
+        })
+
+        const imgUrlArray = await Promise.all(imagexPromiseArray)
+
+const images = imgUrlArray.map((imgUrl) => {
+            return {
+                url: imgUrl,
+            }
+        })
+        const uid = req.user.id
+        console.log(uid)
+        const accommodation = await roomService.findAccommodationById(uid);
+        const hid = accommodation.id
+        console.log(hid)
         const id = uuidv4().replace(/-/g, '');
         const roomData = {
             id,
             name,
+            quantity,
             type,
             price,
-            hostId
+            hostId: hid
         };
 
-        await roomService.createRoom(roomData);
+        const data = await roomService.createRoom(roomData);
+        const rid = data.id
+        const photosdata = await roomService.uploadPhotosRoom({ images, rid });
         res.status(201).json({
             status: 'success',
         });
@@ -54,6 +74,7 @@ exports.updateRoom = async (req, res, next) => {
         const { id } = req.params;
         const {
             name,
+            quantity,
             type,
             price
         } = req.body;
@@ -66,6 +87,7 @@ exports.updateRoom = async (req, res, next) => {
 
         const updatedData = {
             name: name !== undefined ? name : room.name,
+            quantity: quantity !== undefined ? quantity : room.quantity,
             type: type !== undefined ? type : room.type,
             price: price !== undefined ? price : room.price
         };
@@ -86,6 +108,12 @@ exports.updateRoom = async (req, res, next) => {
 exports.deleteRoom = async (req, res, next) => {
     try {
         const { id } = req.params;
+        const room = await roomService.findRoomById(id);
+
+        if (!room) {
+            return next(createError(404, "Room not found"));
+        }
+
         await roomService.deleteRoomById(id);
         res.status(200).json({
             message: "Room deleted successfully"
